Updates CSS to use EM and REM, Cleans up CSS
This commit updates the CSS to use EM and REM units, also implements CSS functions min(), clamp(), fit-content().  This creates a more accessible app that allows the user to zoom in and out, which was not possible with viewports. Adds a CSS variable in App.css '--border-radius'. Cleans up and removes unused code.